# swagger.json
{
	"basePath": "/",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"Name": {
			"properties": {
				"first": {
					"type": "string"
				},
				"last": {
					"type": "string",
					"x-nullable": true
				}
			},
			"required": [
				"first"
			],
			"type": "object"
		},
		"getUserResponseSuccess": {
			"properties": {
				"id": {
					"type": "string"
				},
				"email": {
					"type": "string"
				},
				"name": {
					"$ref": "#/definitions/Name"
				},
				"status": {
					"type": "string",
					"enum": [
						"Happy",
						"Sad"
					],
					"x-nullable": true
				},
				"phoneNumbers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"id",
				"email",
				"name",
				"phoneNumbers"
			],
			"type": "object"
		},
		"ErrorResponse": {
			"properties": {
				"code": {
					"type": "string"
				},
				"message": {
					"type": "string"
				}
			},
			"required": [
				"code",
				"message"
			],
			"type": "object"
		},
		"createUserRequest": {
			"properties": {
				"email": {
					"type": "string"
				},
				"name": {
					"$ref": "#/definitions/Name"
				},
				"phoneNumbers": {
					"type": "array",
					"items": {
						"type": "string"
					}
				}
			},
			"required": [
				"email",
				"name",
				"phoneNumbers"
			],
			"type": "object"
		},
		"createUserResponseSuccess": {
			"properties": {
				"sucecess": {
					"type": "boolean"
				},
				"data": {
					"$ref": "#/definitions/createUserRequest"
				}
			},
			"required": [
				"sucecess",
				"data"
			],
			"type": "object"
		},
		"updateUserRequestBody": {
			"properties": {
				"id": {
					"type": "string"
				},
				"email": {
					"type": "string"
				}
			},
			"required": [
				"id",
				"email"
			],
			"type": "object"
		},
		"updateUserResponse": {
			"properties": {
				"success": {
					"type": "boolean"
				},
				"data": {
					"$ref": "#/definitions/updateUserRequestBody"
				},
				"userId": {
					"type": "number",
					"format": "double"
				}
			},
			"required": [
				"success",
				"data",
				"userId"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "microservice-boilerplate",
		"version": "1.0.0",
		"description": "Simple microservice boiler plate with Node.JS & Typescript",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/users/{id}": {
			"get": {
				"operationId": "GetUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/getUserResponseSuccess"
						}
					},
					"401": {
						"description": "Not Authorized",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					},
					"404": {
						"description": "Not Found",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"description": "Unique user id which is an uuid",
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string",
						"minLength": 24,
						"maxLength": 30
					}
				]
			}
		},
		"/users": {
			"post": {
				"operationId": "CreateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"201": {
						"description": "Created",
						"schema": {
							"$ref": "#/definitions/createUserResponseSuccess"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "requestBody",
						"required": true,
						"schema": {
							"$ref": "#/definitions/createUserRequest"
						}
					}
				]
			}
		},
		"/users/update": {
			"put": {
				"operationId": "UpdateUser",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/updateUserResponse"
						}
					},
					"204": {
						"description": "some invalid data",
						"schema": {
							"$ref": "#/definitions/ErrorResponse"
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "body",
						"name": "data",
						"required": true,
						"schema": {
							"$ref": "#/definitions/updateUserRequestBody"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {}
}